refactor: move kunjungan to separate page
