+++ pages/admin/index.vue
@@ -4,12 +4,9 @@ import type {
   QueryData,
   RealtimePostgresChangesPayload,
 } from "@supabase/supabase-js"
-import type { KunjunganSearchArgs, Peminjaman, PeminjamanDetail } from "@/types"
-import { formatDate } from "#imports"
+import type { Peminjaman, PeminjamanDetail } from "@/types"
 import DataTable from "primevue/datatable"
 import Column from "primevue/column"
-import DatePicker from "primevue/datepicker"
-import FloatLabel from "primevue/floatlabel"
 import Toast from "primevue/toast"
 import { ConfirmDialog } from "primevue"
 
@@ -17,7 +14,6 @@ import { getPeminjamanData } from "@/lib/peminjaman"
 import { useConfirm } from "primevue/useconfirm"
 import { useToast } from "primevue/usetoast"
 import type { Database } from "~/types/database.types.ts"
-import { searchKunjungans } from "~/utils"
 
 useHead({
   title: "Admin",
@@ -172,41 +168,6 @@ async function konfirmasiPengembalian(dataPeminjaman: Peminjaman) {
   }
 }
 
-const kunjunganSearchFor = ref<KunjunganSearchArgs>({
-  timestamp_range: [],
-})
-
-const { data: kunjungans } = await useAsyncData(async () => {
-  const { data, error } = await searchKunjungans(kunjunganSearchFor.value)
-
-  if (error) {
-    toast.add({
-      severity: "error",
-      summary: "Gagal mengambil data",
-      detail: "Gagal mengambil data kunjungan, silahkan coba lagi.",
-    })
-    throw error
-  }
-  return data
-})
-
-async function handleSearchKunjungans() {
-  const { data, error } = await searchKunjungans(kunjunganSearchFor.value)
-
-  if (error) {
-    console.error(error)
-
-    toast.add({
-      severity: "error",
-      summary: "Gagal mencari data peminjaman",
-      detail: "gagal mencari data peminjaman, silahkan coba lagi.",
-      life: 10000,
-    })
-  }
-
-  kunjungans.value = data
-}
-
 async function insertPeminjamandata(payload: RealtimePostgresChangesPayload<Peminjaman>) {
   try {
     const { data, error } = await _peminjamanQuery.eq("id", (payload.new as Peminjaman).id).single()
@@ -331,63 +292,6 @@ onUnmounted(() => {
         <AdminInfoChip to="buku" :data="counts?.buku_count" label="Buku tersedia" />
         <AdminInfoChip to="pengguna" :data="counts?.pengguna_count" label="Pengguna aktif" />
       </ul>
-    </section>
-
-    <section v-if="kunjungans" class="main-section col-span-full">
-      <h2 class="leading-relaxed mb-4">Riwayat Kunjungan</h2>
-
-      <form class="flex gap-4 py-4" @submit.prevent="handleSearchKunjungans()">
-        <FloatLabel>
-          <DatePicker
-            v-model="kunjunganSearchFor.timestamp_range[0]"
-            input-id="start-date"
-            show-button-bar
-            :max-date="new Date()"
-          />
-          <label for="start-date">Tanggal awal</label>
-        </FloatLabel>
-
-        <FloatLabel>
-          <DatePicker
-            v-model="kunjunganSearchFor.timestamp_range[1]"
-            input-id="end-date"
-            show-button-bar
-            :max-date="new Date()"
-          />
-
-          <label for="start-date">Tanggal akhir</label>
-        </FloatLabel>
-
-        <CTA type="submit" label="filter" class="ml-auto" fill />
-      </form>
-
-      <DataTable :value="kunjungans" striped-rows paginator :rows="10">
-        <template #header>
-          <p>Menampilkan {{ kunjungans.length }} kunjungan.</p>
-        </template>
-
-        <Column header="No">
-          <template #body="slotProps">
-            {{ kunjungans.indexOf(slotProps.data) + 1 }}
-          </template>
-        </Column>
-        <Column field="pengguna.nama" header="Pengguna" />
-        <Column header="Waktu">
-          <template #body="slotProps">
-            {{
-              formatDate(new Date(slotProps.data.timestamp), {
-                dateStyle: "long",
-                timeStyle: "short",
-              })
-            }}
-          </template>
-        </Column>
-        <Column header="Status">
-          <template #body="slotProps">
-            {{ slotProps.data.event.replace("_", " ") }}
-          </template>
-        </Column>
-      </DataTable>
     </section>
   </div>
